+++ apps/src/componentLibrary/fontAwesomeV6Icon/FontAwesomeV6Icon.tsx
@@ -13,6 +13,12 @@ export interface FontAwesomeV6IconProps extends AriaAttributes {
    *  * thin - 100
    * */
   iconStyle?: 'solid' | 'regular' | 'light' | 'thin';
+  /** Icon family
+   *    * brands - for FA brand icons
+   *    * duotone - for FA duotone icons
+   *    * kit - for our custom FA kit icons
+   * */
+  iconFamily?: 'brands' | 'duotone' | 'kit';
   /** Icon name */
   iconName: string;
   /** FontAwesome V6 Animation type to use (use it if we want/need to animate icon)*/
@@ -50,7 +56,8 @@ export interface FontAwesomeV6IconProps extends AriaAttributes {
  * Can can be used in any component in/out of the scope of Design System.
  */
 const FontAwesomeV6Icon: React.FunctionComponent<FontAwesomeV6IconProps> = ({
-  iconStyle = 'solid',
+  iconStyle,
+  iconFamily,
   iconName,
   className,
   title,
@@ -63,8 +70,11 @@ const FontAwesomeV6Icon: React.FunctionComponent<FontAwesomeV6IconProps> = ({
     <i
       data-testid="font-awesome-v6-icon"
       className={classNames(
+        iconFamily && `fa-${iconFamily}`,
         iconStyle && `fa-${iconStyle}`,
         iconName && `fa-${iconName}`,
+        // Default icon style is solid, but only when no iconFamily prop is provided
+        !iconFamily && !iconStyle && 'fa-solid',
         animationType && `fa-${animationType}`,
         className
       )}