+++ apps/src/lab2/views/components/versionHistory/VersionHistoryButton.tsx
@@ -90,11 +90,13 @@ const VersionHistoryButton: React.FunctionComponent<VersionHistoryProps> = ({
       <Button
         isIconOnly
         icon={{iconStyle: 'solid', iconName: 'history'}}
-        color={'black'}
+        color={'white'}
         onClick={toggleVersionHistory}
         ariaLabel={commonI18n.versionHistory_header()}
         size={'xs'}
         disabled={buttonDisabled}
+        type={'tertiary'}
+        className={moduleStyles.versionHistoryButton}
       />
       {(loading || loadError) && (
         <div className={moduleStyles.versionHistoryDropdown} ref={menuRef}>